requirements.txt
django
djangorestframework
drf-yasg
django-dataclass-autoserialize
